+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import Banner from "../components/banner/Banner";
 import Footer from "../components/footer/Footer";
 import GlobalStyle from "../GlobalStyle";
-import Main from "../components/Main";
+import Main from "../components/main/Main";
 
 function Home() {
   return (
