+++ src/pages/PostPage.jsx
@@ -1,5 +1,5 @@
 import Banner from "../components/banner/Banner";
-import PostView from "../components/PostView";
+import PostView from "../components/postView/PostView";
 import { useParams } from "react-router-dom";
 
 export function PostPage() {
